Remove redundant functions
