+++ public/app/features/dashboard/state/PanelModel.ts
@@ -22,11 +22,11 @@ import {
   getNextRefId,
   VariableInterpolation,
   builtInVariables,
+  safeStringifyValue,
 } from '@grafana/data';
 import { getTemplateSrv, RefreshEvent } from '@grafana/runtime';
 import { LibraryPanel, LibraryPanelRef } from '@grafana/schema';
 import config from 'app/core/config';
-import { safeStringifyValue } from 'app/core/utils/explore';
 import { QueryGroupOptions } from 'app/types';
 import {
   PanelOptionsChangedEvent,
