+++ public/app/features/explore/CorrelationHelper.tsx
@@ -3,7 +3,12 @@ import { useState, useEffect, useId } from 'react';
 import { useForm } from 'react-hook-form';
 import { useAsync } from 'react-use';
 
-import { DataLinkTransformationConfig, ExploreCorrelationHelperData, GrafanaTheme2 } from '@grafana/data';
+import {
+  DataLinkTransformationConfig,
+  ExploreCorrelationHelperData,
+  GrafanaTheme2,
+  getTransformationVars,
+} from '@grafana/data';
 import {
   Collapse,
   Alert,
@@ -20,7 +25,6 @@ import {
 } from '@grafana/ui';
 import { useDispatch, useSelector } from 'app/types';
 
-import { getTransformationVars } from '../correlations/transformations';
 import { generateDefaultLabel } from '../correlations/utils';
 
 import { CorrelationTransformationAddModal } from './CorrelationTransformationAddModal';
