+++ public/app/features/explore/CorrelationTransformationAddModal.tsx
@@ -3,7 +3,7 @@ import { useId, useState, useMemo, useEffect } from 'react';
 import Highlighter from 'react-highlight-words';
 import { useForm, Controller } from 'react-hook-form';
 
-import { DataLinkTransformationConfig, ScopedVars } from '@grafana/data';
+import { DataLinkTransformationConfig, ScopedVars, getTransformationVars } from '@grafana/data';
 import { Button, Field, Icon, Input, Label, Modal, Select, Tooltip, Stack } from '@grafana/ui';
 
 import {
@@ -11,7 +11,6 @@ import {
   getTransformOptions,
   TransformationFieldDetails,
 } from '../correlations/Forms/types';
-import { getTransformationVars } from '../correlations/transformations';
 
 interface CorrelationTransformationAddModalProps {
   onCancel: () => void;
