+++ public/app/features/variables/inspect/utils.ts
@@ -1,9 +1,8 @@
-import { DataLinkBuiltInVars } from '@grafana/data';
+import { DataLinkBuiltInVars, safeStringifyValue } from '@grafana/data';
 import { Graph } from 'app/core/utils/dag';
 import { mapSet } from 'app/core/utils/set';
 import { stringifyPanelModel } from 'app/features/dashboard/state/PanelModel';
 
-import { safeStringifyValue } from '../../../core/utils/explore';
 import { DashboardModel, PanelModel } from '../../dashboard/state';
 import { variableAdapters } from '../adapters';
 import { isAdHoc } from '../guard';
